fix showing duplicate zero APR
