+++ src/components/PositionsTable/index.ts
@@ -1,2 +1,3 @@
+export * from '../Table/RenderApr'
 export * from './PositionsTable'
 export { default } from './PositionsTable'
