+++ src/components/Table/index.ts
@@ -1,1 +1,2 @@
-export * from "./Table";
+export * from './RenderApr'
+export * from './Table'
